Add project tree to view
