+++ rider-editor/src/ui/file/editor_file_token.rs
@@ -317,8 +317,14 @@ mod tests {
 
     #[cfg_attr(tarpaulin, skip)]
     impl<'l> Renderer for RendererMock<'l> {
+        #[cfg_attr(tarpaulin, skip)]
         fn load_font(&mut self, _details: FontDetails) -> Rc<Font> {
             unimplemented!("load_font")
+        }
+
+        #[cfg_attr(tarpaulin, skip)]
+        fn load_image(&mut self, _path: String) -> Result<Rc<Texture>, String> {
+            unimplemented!()
         }
 
         fn load_text_tex(
